S1 SMPS Example: Use `MPCM_ensureYMWriteReady` before YM writes for extra safety.
The odds of accessing YM when Mega PCM writes to non-DAC registers are *extremely low*, as it only touches other YM registers once when starting/stopping playback. And to break DAC output, M68K needs to "steal" last access YM register in-between instructions that disable FM6, and this needs to be way outside of VBlank, when the M68K sound driver should be done. It's next to impossible, right? So I thought, but someone reported this issue at least once and those "1 in a million chances" unlucky timings were unbelievably consistent.

Loosing extra cycles on every YM write for a "1 in a million chances" bug sucks, but it's a high price we have to pay for reliability.